[test] Fixed  prepare_ui method
- We added two try/except blocks :
* one around the domain evaluation to make it more robust.
* one around the field retrival, in case a field is present in the view, but not in created_model (for instance, a field from a related model)
- In the search of possible values, we use a copy of the domain.
This is because odoo modifies the domain when we call search. This causes weird bugs where domains of other fields were modified.
- We added the uid in the context before calling safe_eval, which brings the domain search closer to the real behaviour.

This is necessary when there are fields that are present in the view but
not in the created_model (e.g. field of a related model that appears on
the view)