Change mechanism to set a date on stock.scrap
date_expected field has been removed from stock.scrap.
But stock.move.line date is now writable.
Therefore we propose a new way of setting the date of a scrap and its
move (to be able to get a correct traceability). The user has to modify
the date of the move line and then use this action.

We could discuss the relevance of keeping the action and not making it a
depends on move_id.stock_move_line.date.