Added script import of JQuery lib for script snippet
Since the module's assets (including the JQuery library) are only loaded after the views and templates have been parsed, it means that the code snippet for handling deletion of administrative documents doesn't recoignize JQuery symbols (such as $).
For this reason, we call the JQuery JS library file. While we could call the whole front-end asset bundle, it might slow down the loading, and so it's better to only load the necessary file.